Done doing Subcomponent
