+++ src/app/add-tournament/add-tournament.component.html
@@ -1,0 +1,78 @@
+<div id="Tournament" class="container-fluid">
+    <form method="post">
+      <div class="row">
+        <div class="col-sm-2"></div>
+        <div class="col-sm-4">
+          <br><br>
+          <input class="form-control" required="true" placeholder="TOURNAMENT NAME" #tournyName />
+          <br>
+          <textarea class="form-control" id="comments" name="comments" placeholder="TOURNAMENT DESCRIPTION" 
+          rows="7" #tournyDesc required="true"></textarea>
+        </div>
+        <div class="col-sm-1">
+          <br><br>  
+            <P class="text-left con_name">GAME TYPE:</P><br><br>
+            <P class="text-left con_name">GAME FORMAT:</P><br>
+            <P class="text-left con_name">NO OF PLAYERS:</P><br><br>
+            <P class="text-left con_name">START DATE:</P>
+        </div>    
+        <div class="col-sm-3">
+          <br>
+          <P class="text-center con_name"><strong>GAME INFORMATION </strong></P>
+          <input class="form-control" placeholder="GAME TYPE" required="true" #tournyType/><br>
+          <input class="form-control" placeholder="GAME FORMAT" required="true" #tournyFormat/><br>
+          <input class="form-control" type="number" placeholder="NO OF PLAYERS (MAX: 250)" #tournyNoPlayers/><br>
+          <input id="startDate" class="form-control con_name" type="date" required="true" #tournyDate/><br>
+          <div class="col-auto ">
+            <button class="btn btn-danger con_name" type="submit" (click)="addNewItem(tournyName.value,tournyDesc.value 
+            + ' [GT:' + tournyType.value + ',GF:' + tournyFormat.value + ',NP:' + tournyNoPlayers.value + ' players]',tournyDate.value)">
+            Add Tournament
+            </button>
+          </div>
+        </div>
+        <div class="col-sm-2"></div>
+      </div>
+    </form>
+    
+    <div class="row">
+      <div class="col-sm-2"></div>
+      <div class="col-sm-8">
+        <br><br>
+        <table class="table table-striped table-bordered table-sm">
+          <thead>
+            <tr class="text-center con_name">
+              <th>Tournament No</th>
+              <th>Tournament Name</th>
+              <th>Tournament Description</th>
+              <th>Start Date</th>
+              <th>Delete Tournament</th>
+            </tr>
+          </thead>
+          <tbody class="tab_content">
+            <tr *ngFor="let item of getItems; let i = index">
+              <td class="centerItem">{{  i+1 }}</td>
+              <td>{{ item.task }}</td>
+              <td>{{ item.task2 }}</td>
+              <td class="centerItem">{{ item.startDate }}</td>
+              <td class="centerItem"><input type="checkbox" [(ngModel)]="item.isComplete"/>
+                <p [ngSwitch]="item.isComplete">
+                  <span *ngSwitchCase="true"></span>
+                  <span *ngSwitchDefault></span>
+                </p> 
+              </td>
+            </tr>
+          </tbody>
+        </table>
+      </div>
+  
+      <!-- <div class="bg-secondary text-white text-left p-2">
+        <div class="form-check">
+          <input class="form-check-input" type="checkbox" [(ngModel)]="showComplete" />
+          <label class="form-check-label" >
+            Show Completed Tasks
+          </label>
+        </div>
+      </div> -->
+  
+    </div>
+</div>
