+++ src/app/app.component.html
@@ -1,145 +1,3 @@
-
-
-<style>
-  :host {
-    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
-    font-size: 14px;
-    color: #333;
-    box-sizing: border-box;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-  }
-
-  h1,
-  h2,
-  h3,
-  h4,
-  h5,
-  h6 {
-    margin: 8px 0;
-  }
-
-  p {
-    margin: 0;
-  }
-
-  .spacer {
-    flex: 1;
-  }
-
-  .toolbar {
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    height: 60px;
-    display: flex;
-    align-items: center;
-    background-color: #fc707c;
-    color: white;
-    font-weight: 600;
-  }
-
-  .toolbar img {
-    margin: 0 16px;
-  }
-
-  .toolbar #twitter-logo {
-    height: 40px;
-    margin: 0 8px;
-  }
-
-  .toolbar #youtube-logo {
-    height: 40px;
-    margin: 0 16px;
-  }
-
-  .toolbar #twitter-logo:hover,
-  .toolbar #youtube-logo:hover {
-    opacity: 0.8;
-  }
-</style>
-
-<div id="Tournament" class="container-fluid">
-  <form method="post">
-    <div class="row">
-      <div class="col-sm-2"></div>
-      <div class="col-sm-4">
-        <br><br>
-        <input class="form-control" required="true" placeholder="TOURNAMENT NAME" #tournyName />
-        <br>
-        <textarea class="form-control" id="comments" name="comments" placeholder="TOURNAMENT DESCRIPTION" 
-        rows="7" #tournyDesc required="true"></textarea>
-      </div>
-      <div class="col-sm-1">
-        <br><br>  
-          <P class="text-left con_name">GAME TYPE:</P><br><br>
-          <P class="text-left con_name">GAME FORMAT:</P><br>
-          <P class="text-left con_name">NO OF PLAYERS:</P><br><br>
-          <P class="text-left con_name">START DATE:</P>
-      </div>    
-      <div class="col-sm-3">
-        <br>
-        <P class="text-center con_name"><strong>GAME INFORMATION </strong></P>
-        <input class="form-control" placeholder="GAME TYPE" required="true" #tournyType/><br>
-        <input class="form-control" placeholder="GAME FORMAT" required="true" #tournyFormat/><br>
-        <input class="form-control" type="number" placeholder="NO OF PLAYERS (MAX: 250)" #tournyNoPlayers/><br>
-        <input id="startDate" class="form-control con_name" type="date" required="true" #tournyDate/><br>
-        <div class="col-auto ">
-          <button class="btn btn-danger con_name" type="submit" (click)="addNewItem(tournyName.value,tournyDesc.value 
-          + ' [GT:' + tournyType.value + ',GF:' + tournyFormat.value + ',NP:' + tournyNoPlayers.value + ' players]',tournyDate.value)">
-          Add Tournament
-          </button>
-        </div>
-      </div>
-      <div class="col-sm-2"></div>
-    </div>
-  </form>
-  
-  <div class="row">
-    <div class="col-sm-2"></div>
-    <div class="col-sm-8">
-      <br><br>
-      <table class="table table-striped table-bordered table-sm">
-        <thead>
-          <tr class="text-center con_name">
-            <th>Tournament Id</th>
-            <th>Tournament Name</th>
-            <th>Tournament Description</th>
-            <th>Start Date</th>
-            <th>Delete Tournament</th>
-          </tr>
-        </thead>
-        <tbody class="tab_content">
-          <tr *ngFor="let item of getItems; let i = index">
-            <td class="centerItem">{{ "T_ID" + (i+1) }}</td>
-            <td>{{ item.task }}</td>
-            <td>{{ item.task2 }}</td>
-            <td class="centerItem">{{ item.startDate }}</td>
-            <td class="centerItem"><input type="checkbox" [(ngModel)]="item.isComplete"/>
-              <p [ngSwitch]="item.isComplete">
-                <span *ngSwitchCase="true"></span>
-                <span *ngSwitchDefault></span>
-              </p> 
-            </td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
-
-    <!-- <div class="bg-secondary text-white text-left p-2">
-      <div class="form-check">
-        <input class="form-check-input" type="checkbox" [(ngModel)]="showComplete" />
-        <label class="form-check-label" >
-          Show Completed Tasks
-        </label>
-      </div>
-    </div> -->
-
-  </div>
-</div> 
-  
-  
-
+<app-add-tournament></app-add-tournament>
