+++ src/app/app.module.ts
@@ -3,10 +3,12 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from "@angular/forms";
 
 import { AppComponent } from './app.component';
+import { AddTournamentComponent } from './add-tournament/add-tournament.component';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    AddTournamentComponent
   ],
   imports: [
     BrowserModule,FormsModule,
